created base test class and reorganized code
