+++ adjudicator/Card_State_Map.txt
