+++ adjudicator/README.md
@@ -1,0 +1,5 @@
+# Hearts Adjudicator
+
+## Documentation
+
+## Unit Testing
