+++ adjudicator/tests/test_state.py
@@ -1,0 +1,8 @@
+import unittest
+from ..state import State
+
+
+class TestState(unittest.TestCase):
+
+    def setUp(self):
+        self.state = State()
